Renamed k_fold_cv to bootstrap_cv; fixed k_fold_cv breaking on certain models; reduced iterations on linear models
The implementation of k_fold_cv was actually a bootstrapping method, not the 'real' K Fold method. I'm now using function pointers in the parameter to use the model's constructor arguments when running bootstrap_cv. Run time on linear models was very long due to them having 100,000 iterations each. Reduced this to 1,000 and upped the learning rate from 0.001 to 0.01.